refactor(web): subfolder the marketing, admin and agent-dock features
