+++ apps/web/src/components/features/admin/index.ts
@@ -1,7 +1,7 @@
-export { AddBoardButton } from "./add-board-button";
-export { AdminBoardsTable } from "./admin-boards-table";
-export { AdminListingsTable } from "./admin-listings-table";
 export { AdminPagination } from "./admin-pagination";
 export { AdminSearchField } from "./admin-search-field";
 export { AdminTabs } from "./admin-tabs";
-export { AdminUsersTable } from "./admin-users-table";
+export { AddBoardButton } from "./boards/add-board-button";
+export { AdminBoardsTable } from "./boards/boards-table";
+export { AdminListingsTable } from "./listings/listings-table";
+export { AdminUsersTable } from "./users/users-table";
