+++ apps/web/src/components/features/agent-dock/update/manual-steps.tsx
