+++ apps/web/src/components/features/agent-dock/update/use-latest-release.ts
