+++ apps/web/src/components/features/marketing/index.ts
@@ -1,14 +1,14 @@
-export * from "./board-strip";
-export * from "./campaign-types";
-export * from "./cta-band";
-export * from "./faq";
-export * from "./faq-items";
-export * from "./hero";
-export * from "./how-it-works";
-export * from "./live-jobs-strip";
 export * from "./marketing-footer";
 export * from "./marketing-nav";
-export * from "./privacy-grid";
-export * from "./product-tour";
 export * from "./section";
 export * from "./section-eyebrow";
+export * from "./sections/board-strip";
+export * from "./sections/campaign-types";
+export * from "./sections/cta-band";
+export * from "./sections/faq";
+export * from "./sections/faq-items";
+export * from "./sections/hero";
+export * from "./sections/how-it-works";
+export * from "./sections/live-jobs-strip";
+export * from "./sections/privacy-grid";
+export * from "./sections/product-tour";
