+++ apps/web/src/components/features/marketing/sections/board-strip.tsx
